Refactor to use variable wrappers
